+++ src/main/java/com/example/nyeon/auth/config/AuthorizationServerConfig.java
@@ -1,5 +1,6 @@
 package com.example.nyeon.auth.config;
 
+import com.example.nyeon.auth.sociallogin.JWTCookieSecurityContextRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -8,18 +9,26 @@ import org.springframework.http.MediaType;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.jwt.JwtEncoder;
 import org.springframework.security.oauth2.server.authorization.config.annotation.web.configuration.OAuth2AuthorizationServerConfiguration;
 import org.springframework.security.oauth2.server.authorization.config.annotation.web.configurers.OAuth2AuthorizationServerConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 import org.springframework.security.web.context.SecurityContextRepository;
+import org.springframework.security.web.savedrequest.CookieRequestCache;
+import org.springframework.security.web.savedrequest.RequestCache;
 import org.springframework.security.web.util.matcher.MediaTypeRequestMatcher;
 
 @Configuration
 @EnableWebSecurity
 public class AuthorizationServerConfig {
     @Autowired
-    private SecurityContextRepository securityContextRepository;
+    private JwtEncoder jwtEncoder;
+
+    @Autowired
+    private JwtDecoder jwtDecoder;
 
     @Bean
     @Order(1)
@@ -40,9 +49,23 @@ public class AuthorizationServerConfig {
                 .oauth2ResourceServer((oauth2) -> oauth2.jwt(Customizer.withDefaults()))
                 // defaultSecurityFilterChain과 SecurityContext를 공유하기 위해 따로 설정해 주어야 한다.
                 .securityContext((securityContext) -> securityContext
-                        .securityContextRepository(securityContextRepository)
+                        .securityContextRepository(securityContextRepository())
+                ).requestCache((requestCache) -> requestCache
+                        .requestCache(cookieRequestCache()))
+                .sessionManagement(session -> session
+                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 );
         // @formatter:on
         return http.build();
+    }
+
+    @Bean
+    public RequestCache cookieRequestCache() {
+        return new CookieRequestCache();
+    }
+
+    @Bean
+    public SecurityContextRepository securityContextRepository() {
+        return new JWTCookieSecurityContextRepository(jwtEncoder, jwtDecoder);
     }
 }
